DataTables Api implemented for product / delete for product now taking place in the index view
Known issues:

Issue #1:
Listings do not update with new products and animals until the ProductListings or AnimalListings view is accessed. This means that if an animal or product is created then deleted before the listing views are clicked on, we get a null exception.

Issue#2: Photos are not being saved properly, when application closes and reopens, new entries photos are not found.

Issue #3: file upload does not submit as a string, and therefore cannot make the string property for PhotoUrl required

Issue #4: tinyaMCE richTextEditor submits with html tags